File Types

+++ app/jobs/pageflow/upload_file_to_s3_job.rb
@@ -4,13 +4,13 @@ module Pageflow
 
     extend StateMachineJob
 
-    def self.perform_with_result(file, options)
+    def self.perform_with_result(file, options = {})
       if file && file.attachment_on_filesystem.file?
         if File.exists?(file.attachment_on_filesystem.path)
           file.attachment_on_s3 = file.attachment_on_filesystem
           file.save!
 
-          file.attachment_on_filesystem.destroy
+          file.attachment_on_filesystem.destroy unless file.keep_on_filesystem_after_upload_to_s3?
           :ok
         else
           logger.info "#{file.class.name} #{file.id} not yet transfered to instance."
